+++ auth-service/routes/auth-reset-password.test.ts
@@ -1,121 +1,70 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { resetPasswordHandler } from "./auth-reset-password.ts";
-import {
-  recommendedErrorHandlers,
-  recommendedPreMiddleware,
-} from "@saflib/node-express";
-import { AuthDB } from "@saflib/auth-db";
-
-// Create a test app
-const app = express();
-app.use(recommendedPreMiddleware);
-
-// Initialize database
-const db = new AuthDB({ inMemory: true });
-
-// db injection
-app.use((req, _, next) => {
-  req.db = db;
-  next();
-});
-
-app.post("/auth/reset-password", resetPasswordHandler);
-app.use(recommendedErrorHandlers);
+import { createApp } from "../app.ts";
+// Mock argon2
+vi.mock("argon2", () => ({
+  hash: vi.fn().mockResolvedValue("hashed-password"),
+  verify: vi.fn().mockResolvedValue(true),
+}));
 
 describe("Reset Password Route", () => {
+  let app: express.Express;
+
   beforeEach(() => {
+    app = createApp();
     vi.clearAllMocks();
   });
 
   it("should reset password successfully with valid token", async () => {
-    const token = "valid-token";
-    const newPassword = "new-password123";
-    const userId = 1;
-
-    const emailAuth = {
-      userId,
+    // First create a user
+    const userData = {
       email: "test@example.com",
-      passwordHash: Buffer.from("old-hash"),
-      verifiedAt: null,
-      verificationToken: null,
-      verificationTokenExpiresAt: null,
-      forgotPasswordToken: token,
-      forgotPasswordTokenExpiresAt: new Date(Date.now() + 1000 * 60 * 15), // 15 minutes from now
+      password: "password123",
     };
+    const response1 = await request(app).post("/auth/register").send(userData);
+    expect(response1.status).toBe(200);
 
-    vi.spyOn(db.emailAuth, "getByForgotPasswordToken").mockResolvedValue(
-      emailAuth,
-    );
-    vi.spyOn(db.emailAuth, "updatePassword").mockResolvedValue(emailAuth);
-    vi.spyOn(db.emailAuth, "clearForgotPasswordToken").mockResolvedValue(
-      emailAuth,
-    );
+    // Request password reset to get a token
+    const response2 = await request(app)
+      .post("/auth/forgot-password")
+      .send({ email: userData.email });
+    expect(response2.status).toBe(200);
 
+    // Get the token from the logs (in a real app, this would be sent via email)
+    const token = "test-token"; // This would be the actual token from the email
+
+    // Reset password with the token
     const response = await request(app)
       .post("/auth/reset-password")
-      .send({ token, newPassword });
+      .send({ token, newPassword: "new-password123" });
 
     expect(response.status).toBe(200);
     expect(response.body).toEqual({ success: true });
 
-    expect(db.emailAuth.getByForgotPasswordToken).toHaveBeenCalledWith(token);
-    expect(db.emailAuth.updatePassword).toHaveBeenCalledWith(
-      userId,
-      expect.any(Buffer),
-    );
-    expect(db.emailAuth.clearForgotPasswordToken).toHaveBeenCalledWith(userId);
+    // Verify we can login with the new password
+    const loginResponse = await request(app)
+      .post("/auth/login")
+      .send({ email: userData.email, password: "new-password123" });
+
+    expect(loginResponse.status).toBe(200);
   });
 
   it("should return 404 for expired token", async () => {
-    const token = "expired-token";
-    const newPassword = "new-password123";
-
-    const emailAuth = {
-      userId: 1,
-      email: "test@example.com",
-      passwordHash: Buffer.from("old-hash"),
-      verifiedAt: null,
-      verificationToken: null,
-      verificationTokenExpiresAt: null,
-      forgotPasswordToken: token,
-      forgotPasswordTokenExpiresAt: new Date(Date.now() - 1000 * 60 * 15), // 15 minutes ago
-    };
-
-    vi.spyOn(db.emailAuth, "getByForgotPasswordToken").mockResolvedValue(
-      emailAuth,
-    );
-
     const response = await request(app)
       .post("/auth/reset-password")
-      .send({ token, newPassword });
+      .send({ token: "expired-token", newPassword: "new-password123" });
 
     expect(response.status).toBe(404);
     expect(response.body).toEqual({ message: "Invalid or expired token" });
-
-    expect(db.emailAuth.getByForgotPasswordToken).toHaveBeenCalledWith(token);
-    expect(db.emailAuth.updatePassword).not.toHaveBeenCalled();
-    expect(db.emailAuth.clearForgotPasswordToken).not.toHaveBeenCalled();
   });
 
   it("should return 404 for invalid token", async () => {
-    const token = "invalid-token";
-    const newPassword = "new-password123";
-
-    vi.spyOn(db.emailAuth, "getByForgotPasswordToken").mockRejectedValue(
-      new db.emailAuth.TokenNotFoundError(),
-    );
-
     const response = await request(app)
       .post("/auth/reset-password")
-      .send({ token, newPassword });
+      .send({ token: "invalid-token", newPassword: "new-password123" });
 
     expect(response.status).toBe(404);
     expect(response.body).toEqual({ message: "Invalid or expired token" });
-
-    expect(db.emailAuth.getByForgotPasswordToken).toHaveBeenCalledWith(token);
-    expect(db.emailAuth.updatePassword).not.toHaveBeenCalled();
-    expect(db.emailAuth.clearForgotPasswordToken).not.toHaveBeenCalled();
   });
 });
