+++ auth-service/routes/index.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { authRouter } from "./auth.ts";
+import { resendVerificationHandler } from "./auth-resend-verification.ts";
 import { loginHandler } from "./auth-login.ts";
 import { logoutHandler } from "./auth-logout.ts";
 import { registerHandler } from "./auth-register.ts";
@@ -15,5 +15,5 @@ router.use("/auth/logout", logoutHandler);
 router.use("/auth/verify", verifyHandler);
 router.use("/auth/forgot-password", forgotPasswordHandler);
 router.use("/auth/reset-password", resetPasswordHandler);
-
+router.use("/auth/resend-verification", resendVerificationHandler);
 export { router as authRouter };
